requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
psycopg2-binary==2.9.9
pgvector==0.2.4
openai==2.6.1
python-dotenv==1.0.1
fastapi==0.109.0
uvicorn==0.27.0 